backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
python-decouple==3.8
celery==5.4.0
redis==5.2.1
Pillow==11.0.0
